# backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Project Management API's
  description: A complete Project Management Software
  version: 0.0.1
servers:
  - url: http://localhost:8000/api
...
paths:
...
  /organisation/create:
    post:
      tags: 
        - organisation
      summary: Create Organisation
      description: Create Organisation
      operationId: createOrganisation
      security:
        - bearerAuth: []
      requestBody:
        description: Create Organisation
        content:
          application/json:
            schema:
              required:
                - email
                - name
              type: object
              properties:
                email:
                  type: string
                name: 
                  type: string
        required: true
      responses:
        '200':
          description: Successful operation
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...